backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
python-dotenv>=1.0.0
chromadb>=0.5.15
pypdf>=5.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
openai>=1.50.0
